LayoutTests/media/media-source/media-managedmse-poster-expected.txt


RUN(source = new ManagedMediaSource())
RUN(video.src = URL.createObjectURL(source))
EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer(loader.type()))
RUN(sourceBuffer.appendBuffer(loader.initSegment()))
EVENT(update)
RUN(sourceBuffer.timestampOffset = 0.012833333333333334)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(index++)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(index++)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(index++)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(index++)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(index++)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(index++)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(index++)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(index++)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(index++)))
EVENT(update)
RUN(sourceBuffer.appendBuffer(loader.mediaSegment(index++)))
EVENT(update)
EXPECTED (video.currentTime >= '1') OK
